fixup commit for tag 'SIGMOD2009_245_ROX'

+++ pathfinder/compiler/core/core.c
@@ -229,7 +229,7 @@ PFcore_num (long long int i)
  * @return A new core tree node
  */
 PFcnode_t *
-PFcore_dec (double d)
+PFcore_dec (long long int d)
 {
     PFcnode_t *core;
